requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Werkzeug==2.0.1
Pillow==8.3.1
pandas==1.3.0
openpyxl==3.0.7
python-docx==0.8.11
pdf2docx==0.5.6
docx2pdf==0.1.7; platform_system=="Windows"
nbconvert==6.1.0
epub2pdf==0.5.0
pywin32==301; platform_system == "Windows" 